Implemented Role-Based Access Control and Permissions

+++ app/Http/Controllers/CommentController.php
@@ -22,6 +22,9 @@ class CommentController extends Controller
     }
 
     public function destroy(Request $request, Feature $feature,Comment $comment){
+        if($comment->user_id !== Auth::id()){
+            abort(403);
+        }
         $comment->delete();
         return to_route('feature.show', $feature);
     }
